Extract fleet wire primitives and align region validation with Core.
Shared Zod schemas move to fleetWirePrimitives so map wire can reuse them; region constraints now require at least one locator like the Python codec.
